OpenGL1: Added lightmap stage pass to FinishShader.

+++ code/renderergl1/tr_local.h
@@ -287,8 +287,7 @@ typedef struct {
 
 typedef struct {
     qboolean        active;
-    int             index;
-
+    byte            index;
     byte            lightmapStyle;
 
     textureBundle_t bundle[NUM_TEXTURE_BUNDLES];
